backend/users/urls_auth.py
"""Authentication URL patterns (register, login, logout, verify)."""

from django.urls import path
from rest_framework_simplejwt.views import TokenObtainPairView, TokenRefreshView

from .views_auth import LogoutView, MeView, RegisterView, VerifyEmailView

urlpatterns = [
    path("auth/register/", RegisterView.as_view(), name="auth-register"),
    # Canonical JWT endpoints
    path("auth/login/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
    path("auth/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
    # Compatibility aliases for frontend expecting djoser-like paths
    path("auth/jwt/create/", TokenObtainPairView.as_view(), name="jwt_create"),
    path("auth/jwt/refresh/", TokenRefreshView.as_view(), name="jwt_refresh"),
    path("auth/me/", MeView.as_view(), name="auth-me"),
    path("auth/logout/", LogoutView.as_view(), name="auth-logout"),
    path("auth/verify-email/", VerifyEmailView.as_view(), name="auth-verify-email"),
]
